requirements.txt
numpy
scipy
ultralytics
opencv-python
pillow